Make CMake set all capabilities on installation
This commit adds code to the CMakeLists file to enable both cap_sys_ptrace
and cap_net_admin when the program is installed. However it changes from
the previous version in that it only adds the capabilities to the
permitted set, not the effective set. This is so that you can run pwait
as a non-root user and use e.g. the ptrace method without acquiring
cap_net_admin, if you have some reason to think a process with
cap_net_admin would cause problems, and vice versa. (I can't really
think of a situation where this helps anything, but why not.)